requirements.txt
asgiref==3.5.2
astroid==2.12.12
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.5
cryptography==39.0.0
defusedxml==0.7.1
dill==0.3.5.1
dj-rest-auth==5.0.1
Django==4.1.2
django-allauth==0.52.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.0.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.5
python3-openid==3.2.0
pytz==2022.5
requests==2.28.2
requests-oauthlib==1.3.1
sendsay-api-python==0.1.4
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.14
wrapt==1.14.1
